compileWarn

+++ mindie_ref/mindie_llm/atb_models/models/baichuan2/13b/model/paged_attention_quant_model.cpp
@@ -142,7 +142,7 @@ atb::Status PagedAttentionQuantModel::InferShape(const std::vector<atb::TensorDe
     const int64_t outDim = graph_.weightTensors.at(graph_.weightTensors.size() - 1).desc.shape.dims[0];
     outTensorDescs.at(0) = graph_.weightTensors.at(0).desc;
     auto outDimNum = inTensorDescs.at(0).shape.dimNum + 1;
-    for (int i = 0; i < outDimNum - 1; i++) {
+    for (uint i = 0; i < outDimNum - 1; i++) {
         outTensorDescs.at(0).shape.dims[i] = inTensorDescs.at(0).shape.dims[i];
     }
     if (param_.isLmHeadParallel) {
@@ -348,7 +348,7 @@ atb::Status PagedAttentionQuantModel::ParseParam(const std::string &param)
 
 atb::Status PagedAttentionQuantModel::BindParamHostTensor(uint32_t nodeId)
 {
-    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= OPERATION_COUNT_BEFORE_LAYER + param_.layerNum) {
+    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= static_cast<uint32_t>(OPERATION_COUNT_BEFORE_LAYER + param_.layerNum)) {
         return atb::NO_ERROR;
     }
