+++ mindie_ref/mindie_llm/atb_models/models/llama_parallel/model/decoder_model.cpp
@@ -326,7 +326,7 @@ atb::Status DecoderModel::BindParamHostTensor(uint32_t nodeId)
     ATB_LOG(INFO) << "BindParamHostTensor";
     ATB_LOG(INFO) << "nodeId = " << nodeId;
 
-    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= OPERATION_COUNT_BEFORE_LAYER + param_.numHiddenLayers) {
+    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= static_cast<uint32_t>(OPERATION_COUNT_BEFORE_LAYER + param_.numHiddenLayers)) {
         return atb::NO_ERROR;
     }
