+++ mindie_ref/mindie_llm/atb_models/models/starcoder/model/paged_attention_quant_model.cpp
@@ -324,7 +324,7 @@ atb::Status PAQuantModel::ParseParam(const std::string &param)
 atb::Status PAQuantModel::BindParamHostTensor(uint32_t nodeId)
 {
     ATB_LOG(INFO) << "BindParamHostTensor";
-    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= OPERATION_COUNT_BEFORE_LAYER + param_.layerNum) {
+    if (nodeId < OPERATION_COUNT_BEFORE_LAYER || nodeId >= static_cast<uint32_t>(OPERATION_COUNT_BEFORE_LAYER + param_.layerNum)) {
         return atb::NO_ERROR;
     }
     bool isFloatLayer = false;
